+++ pyecore/resources/resource.py
@@ -90,10 +90,9 @@ class ResourceSet(object):
     def can_resolve(self, uri_path, from_resource=None):
         uri_path = Resource.normalize(uri_path)
         fragment = uri_path.rsplit('#', maxsplit=1)
-        if len(fragment) == 2:
+        nb_fragments = len(fragment)
+        if nb_fragments == 2:
             uri_str, fragment = fragment
-        else:
-            return False
         if uri_str in self.resources:
             return True
         start = from_resource.uri.normalize() if from_resource else '.'
@@ -105,11 +104,12 @@ class ResourceSet(object):
         upath = Resource.normalize(uri)
         uri_str, fragment = upath.rsplit('#', maxsplit=1)
         if uri_str in self.resources:
-            return Resource._navigate_from(fragment, self.resources[uri_str])
-        start = from_resource.uri.normalize() if from_resource else '.'
-        apath = path.dirname(start)
-        uri = URI(path.join(apath, uri_str))
-        root = self.resources[uri.normalize()]
+            root = self.resources[uri_str]
+        else:
+            start = from_resource.uri.normalize() if from_resource else '.'
+            apath = path.dirname(start)
+            uri = URI(path.join(apath, uri_str))
+            root = self.resources[uri.normalize()]
         if isinstance(root, Resource):
             root_number, fragment = Resource.extract_rootnum_and_frag(fragment)
             root = root.contents[root_number]
@@ -135,6 +135,8 @@ class URI(object):
     def __split(self):
         if '://' in self.plain:
             self._protocol, rest = self.plain.split('://', maxsplit=1)
+        elif ':/' in self.plain:
+            self._protocol, rest = self.plain.split(':/', maxsplit=1)
         else:
             self._protocol, rest = None, self.plain
         uri_sep = self._uri_split.get(self._protocol, path.sep)
@@ -174,14 +176,17 @@ class URI(object):
         return self.__stream
 
     def normalize(self):
-        return self._uri_norm.get(self.protocol, path.abspath)(self.plain)
+        if self.protocol is not None:
+            return self._uri_norm.get(self.protocol, lambda x: x)(self.plain)
+        return path.abspath(self.plain)
 
-    def relative_from_me(self, uri):
+    def relative_from_me(self, other_uri):
         normalized = path.dirname(self.normalize())
-        other = uri
-        if isinstance(uri, URI):
-            other = uri.normalize()
-        return path.relpath(other, normalized)
+        if isinstance(other_uri, URI):
+            other_normalized = other_uri.normalize()
+            if other_uri.protocol:
+                return other_normalized
+        return path.relpath(other_normalized, normalized)
 
     def apply_relative_from_me(self, relative_path):
         parent_path = path.dirname(self.normalize())
@@ -271,11 +276,27 @@ class Resource(object):
         self.uuid_dict = {}
         self.use_uuid = use_uuid
         self.prefixes = {}
-        self.uri = uri
+        self._uri = uri
         self.resource_set = None
         self._decoders = list(Resource._decoders)
         self.contents = []
         self._resolve_mem = {}
+
+    @property
+    def uri(self):
+        return self._uri
+
+    @uri.setter
+    def uri(self, value):
+        uri = value
+        if isinstance(value, str):
+            uri = URI(value)
+        if self.resource_set:
+            old_uri = self._uri.normalize()
+            resources = self.resource_set.resources
+            resources[uri.normalize()] = resources[old_uri]
+            del resources[old_uri]
+        self._uri = uri
 
     def resolve(self, fragment, resource=None):
         fragment = self.normalize(fragment)